pyproject.toml
[project]
name = "devops-engineer-from-scratch-project-313"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.35",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]
